Adding test cases for reading int32
git-svn-id: http://svn.code.sf.net/p/jmatio/code/trunk@94 70b73bdd-3b7a-4af7-8b34-4a2dea44f45c

+++ JMatIO/prod/com/jmatio/types/MLInt32.java
@@ -138,7 +138,7 @@ public class MLInt32 extends MLNumericArray<Integer>
     }
     public int getBytesAllocated()
     {
-        return Long.SIZE >> 3;
+        return Integer.SIZE >> 3;
     }
     
     public Class<Integer> getStorageClazz()
